Move model Enums to models.py
