+++ pooltool/physics/resolve/stick_ball/__init__.py
@@ -1,22 +1,9 @@
 from typing import Dict, Optional, Type
 
 from pooltool.physics.resolve.stick_ball.core import StickBallCollisionStrategy
+from pooltool.physics.resolve.models import StickBallModel
 from pooltool.physics.resolve.stick_ball.instantaneous_point import InstantaneousPoint
 from pooltool.physics.resolve.types import ModelArgs
-from pooltool.utils.strenum import StrEnum, auto
-
-
-class StickBallModel(StrEnum):
-    """An Enum for different stick-ball collision models
-
-    Attributes:
-        INSTANTANEOUS_POINT:
-            Instantaneous and point-like stick-ball interaction
-            (:class:`InstantaneousPoint`).
-    """
-
-    INSTANTANEOUS_POINT = auto()
-
 
 _stick_ball_models: Dict[StickBallModel, Type[StickBallCollisionStrategy]] = {
     StickBallModel.INSTANTANEOUS_POINT: InstantaneousPoint,
